Added collision check.
Implemented point geometry.
Implemented 'point_inside' method of ConvexPolygon class which check
if point is inside polygon.
Implemented 'collides' method for GamePiece class which checks if the
geometry which represents the GamePiece collides with the geometry which
represents the other GamePiece.

Implemented test for Point class.
Implemented test for 'point_inside' method of ConvexPolygon class.
Implemented test for 'collides' method of GamePiece class.