Added collision logic.
'_resolve_collision' method of GameBoard class checks if a ship
has collided with an asteroid and sets the _gameover flag to True.
'_resolve_collisoin' method of GameBoard class checks if a projectile
has collided with an asteroid. Both the projectile and the asteroid
are subsequently removed from the gameboard.

Test for collision detection added.